solutions/5/guillaume/common.py
import os
import random

import numpy as np
from numpy import linalg as LA

import pandas as pd

import tqdm

try:
    import torch
    import torch.opt
    import torch.nn as n
    import torch_geometric.nn as gn
    import torch.nn.functional as F
    from torch.utils.data import Datas
except:
    pass